refactor: moved all functions for working with entry to a separate file

+++ include/utils.h
@@ -14,7 +14,7 @@ char *kl_strdup(const char *str);
 bool kl_strequal(char *str1, char *str2);
 int kl_strncmp(const char *str1, const char *str2, size_t n);
 int kl_numlen(long long n);
-void print_num(int num);
+void print_num(long long num);
 void print_str(char *str);
 void print_char(char c);
 void print_repeat_char(char c, size_t len);
